Handle review comment
